+++ src/client/mod.rs
@@ -384,6 +384,7 @@ mod parity_contract {
             let _ = $module::Handle::<Class>::relation::<ClassRelationId>;
             let _ = $module::Handle::<Class>::create_relation::<ClassId>;
             let _ = $module::Handle::<Class>::create_relation_with_aliases::<ClassId>;
+            let _ = $module::Handle::<Class>::create_relation_with_options::<ClassId>;
             let _ = $module::Handle::<Class>::delete_relation::<ClassRelationId>;
 
             let _ = $module::Handle::<Object>::related_objects;